Hey, I have noticed that the west germans are missing a unit as a formation support. I bought a Leapord 1 box for west germans and found a unit card called Gebirgspanzerjager zug-1.Gebirgsdivision. It is a squad of x5 Leapord 1 tanks for the west germans, I notoced when lookong through battlescribe, this  squad is nowhere to be found. I THOUGHT i would let you know so you can add it in. I am using the latest version of battlescribe.
One more thing I should add, the warning symbol that comes up with an error if a unti is not allowed to be added to your squad, nver comes up for west germans. I can add 10 squads of Pah helicopters or tornado's with no warning saying this is not allowed. West germans are only allowes x2 squads of Pah helicopters and x1 squad of 4 jets.

I hope this informations helps and thanks for you time.